gcfunction/requirements.txt
numpy == 1.17
torch == 1.5.0
torchvision == 0.6.0
google-cloud-storage == 1.30.0
scipy == 1.5.2
